1月1号1.1.2版本提审

+++ LearnMore/Classes/Profile/Model/LMActBook.h
@@ -11,6 +11,6 @@
 @interface LMActBook : NSObject
 @property (nonatomic, assign) long long typeId;
 @property (copy, nonatomic) NSString *actTitle;
-@property (nonatomic, assign) long actBeginTime;
-@property (nonatomic, assign) long actEndTime;
+@property (nonatomic, assign)  long long actBeginTime;
+@property (nonatomic, assign) long long actEndTime;
 @end
